Script to run BASH commands on the WEB container for DEV/ALPHA
To obtain a SQL dump from environment 'dev', run it like so:

    bin/clam dev 'mysqldump -h $RDS_ADDRESS -u $DB_USERNAME --password=$DB_PASSWORD idseq_dev | gzip -c' | gzip -dc > id_dev.sql

Note how the environment variables are expanded on the remote container,
and how the sql dump is compressed on the remote machine, streamed in
compressed format, and uncompressed on the receiving machine.

This differs from the bin/shell script in that it does not run the command
in a tty but does run it in the correct container environment to enable
the expansion of environment variables and shell pipes.

Make sure to use single-quotes as in the above example.