new hit calling and poll single output file
Handle the new data type with pipeline version >= 1.5 that identidies hits
which map to multiple species within a genus and group them under a
line item "non species-specific hits to genus".

Poll a single output file emitted at the end of a successful batch job,
as opposed to requiring the presence of a number of files.  This leads
to a more robust pipeline in future, and to a more tolerant webapp
in present (able to support pipeline versions before and after 1.5).