Jira-311: fixing phage category filter and other category-related bugs (#1478)
Phage category filter wasn't working properly, in fact doing the opposite of what it should (when Phage was selected, then it would be excluded). Also, whenever any category was removed using the 'X' on the blue pill, phage would be removed. Finally, whenever any category was removed using the 'X' or filters were reset, the state of the MultipleDropdown would silently preserve the old state, which would show up in weird ways the next time an item was selected.

All three of those issues are resolved in this diff.