Fix bug where heatmap broke if filter returned no rows.
Previously, if a filter returned no matching rows, the sample objects wouldn't be passed
back in the request.

SamplesHeatmapView overwrote its current sample ids with whatever sample ids were passed back,
which means that after an empty filter, the sample ids permanently disappeared from
the heatmap view.

To fix this, we make sure SamplesHeatmapView preserves its original sample ids, regardless of what
is passed back in the data request.

We also have the samples_taxon endpoint pass back the samples (minus the taxon list) even if taxon_ids
is empty. This is so that the heatmap view can correctly compute the set of host genomes in the samples
when the page loads.