Apply new styles to ui components and refactor BareDropdown.
Major style changes are reducing the font size in the dropdown menus
and reducing the height of buttons and triggers.

We also make all dropdowns scrollable by default, and add some styling to
the scrollbars.

In order to make styling more consistent, we refactor BareDropdown to
render the Dropdown Items manually in nested Dropdown Menus. (that's
the only way to get a scrollbar inside the Menu)

We also refactor several components usage of BareDropdown so they
can receive the new styling.