Fix bugs caused by metadata schema changes. (#1868)
Adding collection_location back to mosquito since front-end has hard-coded
dependencies on it. This fixes an issue in Heatmap that causes it to crash
when SearchBoxList is clicked.

Mosquito already has collection_long and collection_lat, so this creates a duplication.
However, collection_location will NOT appear in the metadata sidebar, so the duplication will
not be noticeable. This will be temporary until the new metadata types table comes online.

Also re-adding id_method and genus_species to the keys. These two keys are still
part of the vector metadata schema, so removing them was a mistake. This fixes an
issues on phylo tree where an empty string was being displayed as a dropdown option
for id_method and genus_species.