[IDSEQ-901] Fix loss of work issues in Sample Upload Flow. (#2110)
We no longer delete the metadata when the user goes back to the Upload Samples step.

For CSV metadata, we re-validate the metadata when MetadataCSVUpload becomes visible again if we detect
that the samples have changed. For manual metadata, the validate is still done when the user clicks
Continue (as before).

We are also more intelligent about when to disable the header navigation.

Now, we only disable the nav once the user has changed something
(which is tracked via the "onDirty" method). In this case, the user has to
click the "Continue" button to go through the steps again.
This is because (among other reasons) metadata validation only happens at the Upload Metadata step.
We don't want the user to be able to change the samples and then jump straight to the Review step, so
we make them go through the flow again.