Add coverage viz to idseq-web dag template.
Currently gated with pipeline version >= 3.6.

As soon as the coverage viz step gets pushed to idseq_dag master, idseq_web will start
running that pipeline (due to this change).
We plan to test the coverage viz step in staging and prod using admin options before pushing
the coverage viz step to master.

The front-end is further gated on allowed_feature=coverage_viz_enabled and admin-only,
which we will remove in an upcoming PR.