Add ContigDownloader to Coverage Viz.
Add functions to fetch byteranges of the contigs.fasta file.

Refactor the HitGroupViz into a separate component.
Add new Copy Icon.
Fix issue where tooltip was overflowing the window on the right,
and becoming narrower as it did so.

Make tooltips and contig downloader position:fixed instead of
position:absolute, so they'll stay in the correct position if
the user scrolls the page.