Add CoverageIcon for Hover Action and tweak coverage viz styling. (#2178)
Show new Coverage Viz hover action if pipelineVersion >= 3.6. Show old Alignment Viz action for older pipelines.
Refactor HoverActions to pass params in JS instead of via data attributes.

Small tweaks to naming and display of metrics.

No more Popups in the Accession Select dropdown.

Reworked the Coverage Viz header so that the Accession Select is on the left, next to
the Accession Count display.

Added menuClass prop to BareDropdown so we can set a min-width for the dropdown menu.