Show Notification toast after Metadata Upload. (#2193)
* Show Notification toast after Metadata Upload.

Add new ListNotification component for notifications that contain a list of items.
Refactor Accordion so it can be used in more places. This was mainly removing opinionated styling.

Add Success state to NotificationComponent.

Refactor PublicSampleNotification to use ListNotification.

* Rearrange file structure of notifications and toasts.

Notifications are now grouped together in /ui/notifications.
They are components that show info, warnings, errors, and success states.

Toasts describe anything that appears at the top-right and can be dismissed.
Often you will create toasts containing a notification.
Notifications can also be used outside of toasts, for example to display errors in a form.