Heatmap zoom (#2522)
Heatmap zoom# Description

See https://jira.czi.team/browse/IDSEQ-1233

This adds zoom to heatmap so that users can see more or less of a heatmap as they wish. In addition, it adds shrink-to-fit so that all samples of a heatmap are visible on initial load. 

To keep the scrollbars in view on zoom, I changed the heatmap container to have a fixed height. A side effect of this is to lose the semi-fixed page header on scroll. Now scrolling controls the content of the heatmap only, and the content above does not move, so it is effectively the same, except that the black idseq header and other content above the filters does not disappear. This change also fixes the visual issue where the fixed filter bar did not cover the whole width of the screen, leaving awkward patches of heatmap on either side of it. If any of this is not clear enough, please ask for me a demo. 

Lastly, the fixed height container revealed a latent issue where tooltips at the bottom got clipped on hover. I fixed this by extending @MarkAZhang 's fix for the same issue on the right side of page -- the tooltips now show above when close to the bottom. Note that the fix assumes two constant heights of the tooltips. I'm not sure how to get the height of the content before it has rendered. 

## Zoom out
![image](https://user-images.githubusercontent.com/28797/63134742-9dc20200-bf7f-11e9-8a78-39f50bd3617c.png)

## Zoom in
![image](https://user-images.githubusercontent.com/28797/63134751-aadef100-bf7f-11e9-922e-9160e349d687.png)

## Zoom control
![image](https://user-images.githubusercontent.com/28797/63301922-ad4b8e80-c290-11e9-9576-d9e2316c0900.png)


## Flip tooltip at bottom of page
![image](https://user-images.githubusercontent.com/28797/63301878-91e08380-c290-11e9-8592-3030a35a0405.png)

![image](https://user-images.githubusercontent.com/28797/63301881-9311b080-c290-11e9-8dfa-011490193acb.png)

# Notes

* While the text becomes too small to read, the hover tooltip is still at normal size, allowing a user to read the contents of a cell
* The +/- control is the same one used on the map and pipeline viz

# Tests

**Zoom in and out**
Load a heatmap

Click on the plus 
See the heatmap expand up to the max

Click on the minus
See the heatmap shrink
See the heatmap stop shrinking at the smallest size

**Shrink-to-fit**
Shrink the size of your browser to <1000px wide
Load a large heatmap
See the heatmap fit in the width
Enlarge the browser to full screen
Reload
See the heatmap at the default size (1300px)