Fix bug in location dropdown with apply all (#2725)
Fix bug in location dropdown with apply all# Description

This fixes some related bugs discovered by @happyimadesignr  and @oliviabholmes . I had added `currentResult` to the input component state so that the blur handler could validate `currentResult` and produce warnings. However, the "apply to all" function set the input value by a different code path. In consequence, when focusing in and out of an input that had a value "applied" to it, bad things happened. 

The solution is simply to always clear `currentResult` onblur, and not validate when `currentResult` is blank. 

Warnings will only appear in the single input the user is focussed on. Warnings will not appear on other inputs when "apply all" is applied. I think this is consistent with the original intention. 

# Notes

There is another bug I uncovered in testing that is already in prod and probably always has been. If you trigger a warning on a input, then override the value with "apply all" from another input, the warning will *not* go away. 

![image](https://user-images.githubusercontent.com/28797/68247195-22487d80-ffcf-11e9-9fde-d18204c984a4.png)

@MarkAZhang  has proposed me a fix already, which I may add here. 

# Tests

* Upload more than one sample
* Match a location
* Hit apply to all
* Focus in and out of second input
* See value stay the same, no warning