[Release fix] Support reportReady and pipelineRunReportAvailable in case of caching (#2887)
[Release fix] Support reportReady and pipelineRunReportAvailable in case of caching# Description

The variables `pipelineRunReportAvailable` and `report_ready` were consolidated and renamed to `reportReady`, but a version of the code using the old variable names may have been cached in the browser. Since the front-end only checks for the new `reportReady`, this can result in a sample report appearing as failed even if it succeeded.

This change makes it so that so both versions of the variables are supported, and will be removed as part of cleanup when the caches have been cleared.