Fix bugs and edge cases in User Create/Update flows (#2912)
Fix bugs and edge cases in User Create/Update flows### Description
Bundled fixes for a couple of issues and edge cases that were brought up:
(1) Hide the "send activation email" checkbox when you're updating a user (not making a new one).
(2) Clarify the forgot password workflow with a message (we had this message on the Auth0-hosted screen as well).
(3) Show user creation error messages properly.
(4) Add a validation check for email uniqueness (we had it on the index but not in the validator) and disallow uppercase characters.

### Notes
(3) Before we had the pattern for `if @user.save / else`, but that only handles validation checks, so exceptions could be thrown and not properly passed to the frontend. This was causing the frontend page to crash without explanation.

(4) We discovered that Auth0 automatically converts all emails to lowercase and treats them as case-insensitive, even when our input had capital letters (see: [link](https://community.auth0.com/t/creating-a-user-converts-email-to-lowercase/6678) and [link](https://community.auth0.com/t/api-is-converting-username-to-lowercase/7763)). This would make it impossible for the user to log in because of the inconsistency between our field and theirs.

Based on email specs, an email could technically be case-sensitive, but in reality it's very unlikely that an email provider would allow case-sensitive emails (e.g. bob@ as different from Bob@). We could either lowercase emails automatically as well, or warn on capital letters at creation time. I did the latter for user clarity and to be a little safer. I also checked staging/prod and did not find any other emails with capital letters, so this is an edge case.

### Tests
<img width="509" alt="Screen Shot 2019-12-30 at 2 29 36 PM" src="https://user-images.githubusercontent.com/5652739/71605531-58f9ca80-2b1e-11ea-94f8-281298b7e7a1.png">
<img width="494" alt="Screen Shot 2019-12-30 at 3 00 44 PM" src="https://user-images.githubusercontent.com/5652739/71605537-67e07d00-2b1e-11ea-8a2a-55ae9943494d.png">
<img width="520" alt="Screen Shot 2019-12-30 at 3 06 31 PM" src="https://user-images.githubusercontent.com/5652739/71605541-6a42d700-2b1e-11ea-9ca2-ba64122faa99.png">