Add validation to all models, part 1 of 2 (#3074)
Add validation to all models, part 1 of 2# Description

This is a revival of #2478 using a more incremental approach. Apologies for the long description but I feel it's important explain more than usual for such a change.

## Motivation

Historically, idseq-web has used few active record validators and has not consistently restricted the use of `NULL` values in MySQL (which are passed on as `nil` in ruby and `null` in JS). This causes cognitive load on all developers of idseq-web, if not outright bugs. 

For example, recently I added a null host genome to a sample in staging by accident because of a bug. The effect was to crash the data discovery interface for all users who could see the sample. Programmers throughout the app should not have to defend against any value coming from the database being `NULL`. They should be able to safely assume that certain values are always present. 
 
(Note that the uncertainty is compounded by the dynamic languages Ruby and Javascript which means that there are no guarantees anywhere in the stack.)

See also https://guides.rubyonrails.org/active_record_validations.html#why-use-validations-questionmark .

## Background

Six months ago, In an attempt to improve the situation, I found all the columns in our prod database which have all non-NULL values but which will still allow NULLs. See https://czi.quip.com/LWbJAyLuX7fn/2019-08-02-NULL-values-in-database . In #2478, I merely wrote a migration to change columns to `NOT NULL`. I abandoned #2478 because the rollout would have been too risky. Any intermediate state of the database where a `NOT NULL` column would need to hold a `NULL` temporarily would cause a crash. And testing all affected code paths would have been impractical.

## Implementation

This PR leverages active record validators to check for `NULL` values while avoiding any crash in production by making validation depend on the deployment environment. For example:

```
  validates :s3_nt_db_path, presence: true, if: :mass_validation_enabled?

  def mass_validation_enabled?
    Rails.env.development? ||
      Rails.env.test? ||
      AppConfig.find_by(key: AppConfig::ENABLE_MASS_VALIDATION)
  end
```
 The `AppConfig` exists to later enable switching on in staging and production.

Note that the validators in this PR provide considerable value as documentation alone. 

The full change algorithm I followed was:
1. Inspected prod db (as described above)
2. Created list of "no nulls" columns that were not "not null" already
3. Added `validates :attribute, presence: true` to all such columns as model attributes
4. Added extra validation where obvious
* Inclusion in set of constants
* Range of number
5. Added `if: :mass_validation_enabled?` to all new validators
6. Ran all mintiest and rspec
        * Updated fixtures and factories as needed
        * Updated test cases only when no good alternative

# Notes

* To aid my work and future code readers, I standardized the order of lines at the top of models. The order I aimed for is:
  1. Assocs
  2. Validations
  3. Callbacks
  4. Constants
* I discovered some Rails-level associations were missing, most often from some model that `belongs_to` `TaxonLineage` via `taxid`. I added these but commented them out if they caused any unit test failure for lack of time. I feel the associations are important at least as documentation. 
* I found we always use integer in `[0, 1]` instead of Rails and Mysql boolean types. I don't think it is worth the effort to refactor these but I would support changing the pattern for new columns.
* I found some columns which had always the same value in the prod database so it looks like they are not serving any purpose. Similarly, the taxon scoring model model is a singleton after several years in production.
* Annoying mix of `taxid` and `tax_id`. What is preferred?
* Several tables copy data from an associated table without explanation. This violates the normalization of the database (which is nearly complete). What is the point of having `tax_name` and `tax_id` when `tax_name` is derived from `tax_id`?
* A lot of "status" columns that allow `NULL` which seems pretty bad
* A bunch of json columns with different APIs. Why can't we use one of many gems for json attributes? 
* I added some comments on current values for columns that looked like enums but for which it wasn't clear where the possible values are defined. Any pointers would be most welcome to complete the documentation. 
* I got an odd rspec error when running certain files in isolation, such as: `"WebMock::NetConnectNotAllowedError: Real HTTP connections are disabled. Unregistered request"`

# Tests

Ran all mintiest and rspec. See point 6. above. 

Created a project and uploaded a sample in local host.

# Rollout plan

After merging to master, my plan is to: 
1. wait two weeks for any validation errors that happen in developers's localhosts. 
2. turn on `ENABLE_MASS_VALIDATION` in staging
3. ask QA team to exercise all features that create or update objects
4. wait two weeks for any new validation errors that happen in staging
5. turn on `ENABLE_MASS_VALIDATION` in prod
6. react to any new prod validation error by turning off `ENABLE_MASS_VALIDATION` then fixing the issue
7. begin on the second PR which will add `NOT NULL`s to the database schema