[CH-11256] On the report page, when a request with a new background fails, does not report the error and it shows the old background (#342)
# Description
On the report page, when a request with a new background fails, it does not report the error and it shows the old background.

# Notes
Design:
![Screen Shot 2020-09-22 at 2 58 33 PM](https://user-images.githubusercontent.com/51972068/93941833-18616300-fce4-11ea-992d-6f700e6f3e9f.png)

Implementation:
GIF too large to attach here. I have sent in the implementation change in the slack channel: #proj-idseq

# Tests
Tested the implementation on multiple samples, clicked dismiss button, waited for autoClose (12 seconds), ensured that the sample report page and filters worked as expected.