Fix CLI install instructions (#660)
- Do not give the user outdated Homebrew install instructions; instead
  give the user a link to the Homebrew website

- Use pip3 only; do not ask the user to use pip, which defaults to
  Python 2, is deprecated, unsupported, and broken

- Do not ask the user to install Python package using git+https://,
  which implies an undocumented dependency on git and affords no
  pre-release testing of the CLI. Instead, ask the user to install the
  released version of the package from PyPI

Git is generally not suitable for distribution of Python packages to
non-developers. We should always test then release the CLI package on
PyPI when updates are made. A release target was added to the Makefile
in d2b367216434aa9459eca0f7d3f73730811cb22d. IDseq developers should be
added as maintainers to the PyPI package as necessary.

* Update app/assets/src/components/views/CliUserInstructions.jsx

Co-authored-by: Jonathan Sheu <jshoe@users.noreply.github.com>